plataforma/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-hcaptcha-field==1.4.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-vite==3.1.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
idna==3.10
inertia-django==1.2.0
Markdown==3.8.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
